Factor out diagnostic consumer.
