+++ toolchain/check/BUILD
@@ -26,6 +26,7 @@ cc_library(
         "cpp/constant.cpp",
         "cpp/context.cpp",
         "cpp/custom_type_mapping.cpp",
+        "cpp/diagnostic_consumer.cpp",
         "cpp/export.cpp",
         "cpp/generate_ast.cpp",
         "cpp/impl_lookup.cpp",
@@ -90,6 +91,7 @@ cc_library(
         "cpp/constant.h",
         "cpp/context.h",
         "cpp/custom_type_mapping.h",
+        "cpp/diagnostic_consumer.h",
         "cpp/diagnostic_listener.h",
         "cpp/export.h",
         "cpp/generate_ast.h",
