+++ toolchain/check/cpp/diagnostic_listener.h
@@ -5,8 +5,11 @@
 #ifndef CARBON_TOOLCHAIN_CHECK_CPP_DIAGNOSTIC_LISTENER_H_
 #define CARBON_TOOLCHAIN_CHECK_CPP_DIAGNOSTIC_LISTENER_H_
 
+#include <string>
+
 #include "clang/Basic/Diagnostic.h"
-#include "llvm/ADT/StringRef.h"
+#include "clang/Basic/SourceLocation.h"
+#include "llvm/ADT/ArrayRef.h"
 
 namespace Carbon::Check {
 
@@ -19,17 +22,23 @@ class CarbonClangDiagnosticConsumer;
 // itself on destruction.
 class CppDiagnosticListener {
  public:
+  struct Diagnostic {
+    clang::DiagnosticsEngine::Level level;
+    clang::SourceLocation location;
+    std::string message;
+    std::string snippet;
+  };
+
   explicit CppDiagnosticListener(CarbonClangDiagnosticConsumer& consumer);
   virtual ~CppDiagnosticListener();
 
-  // Handles a Clang diagnostic and any associated code snippet.
-  virtual auto HandleDiagnostic(clang::DiagnosticsEngine::Level diag_level,
-                                const clang::Diagnostic& info,
-                                llvm::StringRef message,
-                                llvm::StringRef snippet) -> void = 0;
-
-  // Flushes any buffered diagnostics.
-  virtual auto Flush() -> void {}
+  // Emits a group of buffered diagnostics, which will comprise a single leading
+  // diagnostic -- typically an error / warning / remark -- followed by a
+  // sequence of attached notes. The first diagnostic can also be a note if the
+  // diagnostics engine is flushed in the middle of Clang emitting a diagnostic,
+  // or if Clang emits a "stray" note not attached to any diagnostic. Both would
+  // generally indicate Clang bugs.
+  virtual auto EmitDiagnostics(llvm::ArrayRef<Diagnostic> diags) -> void = 0;
 
  protected:
   auto consumer() -> CarbonClangDiagnosticConsumer& { return *consumer_; }
